+++ cypress/integration/detail.js
@@ -13,7 +13,7 @@ context('detail', () => {
 
         cy.eyesCheckWindow('plex-detail - scrool');
 
-        cy.get('plex-layout plex-help').click();
+        cy.get('plex-layout plex-help').plexButtonIcon('information-variant').click();
 
         cy.eyesCheckWindow('plex-detail - plex-help');