+++ cypress/support/plex.js
@@ -142,7 +142,7 @@ Cypress.Commands.add('plexTextArea', { prevSubject: 'optional' }, (subject, labe
         element = cy.get(`plex-text[${label}] textarea`).first();
     }
     if (text) {
-        element.type(text);
+        element.type(text, { force: true });
     }
     return element.parent().parent().parent();
 });
@@ -269,24 +269,17 @@ Cypress.Commands.add('validationMessage', { prevSubject: true }, (subject, text)
     return cy.wrap(subject).find('div[class="form-control-feedback"]').should('contain', text);
 })
 
-
-/**
- * @decrecated
- */
-Cypress.Commands.add('selectOption', (label, value) => {
-    return cy.get(`plex-select[${label}]`).children().children('.selectize-control').click()
-        .find(`.option[data-value=${value}]`).click({
-            force: true
-        });
+Cypress.Commands.add('tooltip', { prevSubject: true }, (subject, text) => {
+    return cy.wrap(subject).parent().parent().find('.tooltip-inner').should('contain', text);
 });
 
-/**
- * @decrecated
- */
+Cypress.Commands.add('toast', (option, label) => {
 
-Cypress.Commands.add('selectWrite', (label, value) => {
-    return cy.get(`plex-select[${label}] input`).first().type(`${value}{enter}`, {
-        force: true
-    });
-});
+    if (label) {
+        return cy.get(`div[class="simple-notification toast ${option}"]`).contains(label).click();
+    } else {
+        return cy.get(`div[class="simple-notification toast ${option}"]`).click();
+    }
 
+
+})