+++ index.ts
@@ -3,3 +3,4 @@ export * from './src/lib/core/service';
 export * from './src/lib/dropdown/dropdown-item.inteface';
 export * from './src/lib/select/select-event.interface';
 export * from './src/lib/options/options.component';
+export * from './src/lib/text/text.component';