+++ src/demo/app/templates/modulos/mpi/sidebar/detalle/mpi-detalle.component.ts
@@ -16,9 +16,11 @@ export class MpiDetalleComponent implements OnInit {
 
     cerrarSidebar() {
         this.cerrar.emit(this.sidebar);
+        console.log(this.sidebar)
     }
 
     public listadoPaciente: Paciente[];
+    pacientes$: Observable<Paciente[]>;
     paciente$: Observable<Paciente>;
     public items = [
         {
@@ -71,6 +73,8 @@ export class MpiDetalleComponent implements OnInit {
 
     ngOnInit() {
 
+        this.pacientes$ = this.pacienteService.getPacientes();
+
         this.paciente$ = this.route.paramMap.pipe(
             switchMap((params: ParamMap) =>
                 this.pacienteService.getPaciente(params.get('id')))
@@ -79,9 +83,7 @@ export class MpiDetalleComponent implements OnInit {
 
     gotoPacientes(paciente: Paciente) {
         const pacienteId = paciente ? paciente.id : null;
-        this.router.navigate(['/listado-sidebar', { id: pacienteId, foo: 'foo' }]);
+        this.router.navigate(['/listado-sidebar', { id: pacienteId }]);
     }
-
-
 }