+++ src/lib/app/app.component.ts
@@ -1,10 +1,10 @@
-import { Component, OnInit, Input, Renderer2, ViewChild, ViewContainerRef } from '@angular/core';
+import { Component, OnInit, Input, Renderer2, ViewChild, ViewContainerRef, ContentChild } from '@angular/core';
 import { Plex } from './../core/service';
 
 @Component({
     selector: 'plex-app',
     template: ` <!--Navigation Bar-->
-                    <nav class="navbar-inverse fixed-top"  [ngClass]="'bg-' + type">
+                    <nav [hidden]="!plex.navbarVisible" class="navbar-inverse fixed-top bg-{{type}}">
 
                         <div class="navbar-container">
                             <ng-content select="[navIcon]"></ng-content>
@@ -49,18 +49,18 @@ import { Plex } from './../core/service';
                                 </div>
                             </div>
                             <!--Menu-->
-                            <div *ngIf="plex.menu && plex.menu.length" class="action dropdown" [ngClass]="{show: menuOpen}" (click)="toggleMenu(); $event.stopImmediatePropagation();">
+                            <div role="button" *ngIf="plex.menu && plex.menu.length" class="action dropdown" tabindex="1" [ngClass]="{show: menuOpen}" (click)="toggleMenu(); $event.stopImmediatePropagation();">
                                 <i class="mdi mdi-menu"></i>
                                 <ul class="dropdown-menu dropdown-menu-right">
-                                    <li *ngFor="let item of plex.menu">
+                                    <li *ngFor="let item of plex.menu; let i = index">
                                         <!--Item con router asociado-->
                                         <ng-template [ngIf]="!item.divider && item.route">
-                                            <a plexRipples class="dropdown-item" href="#" [routerLink]="item.route" routerLinkActive="active">
+                                            <a plexRipples class="dropdown-item" href="#" tabindex="{{i + 1}}" [routerLink]="item.route" routerLinkActive="active">
                                                 <span *ngIf="item.icon" class="mdi mdi-{{item.icon}}"></span> {{item.label}}</a>
                                         </ng-template>
                                         <!--Item con handler asociado-->
                                         <ng-template [ngIf]="!item.divider && item.handler">
-                                            <a plexRipples class="dropdown-item" href="#" (click)="item.handler($event); false;">
+                                            <a plexRipples class="dropdown-item" href="#" tabindex="{{i + 1}}" (click)="item.handler($event); false;">
                                                 <span *ngIf="item.icon" class="mdi mdi-{{item.icon}}"></span> {{item.label}}</a>
                                         </ng-template>
                                         <!--Divider-->
@@ -82,7 +82,7 @@ import { Plex } from './../core/service';
                 <ng-content select="plex-ribbon"></ng-content>
 
                 <!--Contenedor principal-->
-                <div class="content">
+                <div class="content" [ngClass]="{'nav-top-margin': plex.navbarVisible, 'nav-top-no-margin': !plex.navbarVisible}">
                     <router-outlet></router-outlet>
                 </div>`,
 })
@@ -91,7 +91,7 @@ export class PlexAppComponent implements OnInit {
     // Referencia al DOM para injectar una componente de forma dinámica
     @ViewChild('menuItem', { static: true, read: ViewContainerRef }) viewContainerRef: ViewContainerRef;
 
-    @Input() type: String = 'inverse';
+    @Input() type = 'inverse';
     public loginOpen = false;
     public menuOpen = false;
     public online = true;