+++ src/lib/item-list/heading.component.ts
@@ -3,13 +3,15 @@ import { Component, ChangeDetectorRef, Input } from '@angular/core';
 @Component({
     selector: 'plex-heading',
     template: `
-    <div class="item-list-heading" [class.sticky]="sticky" [class.has-icon]="hasIcon" [class.has-checkbox]="hasCheckbox">
-        <b *ngIf="hasCheckbox"></b>
-        <b *ngIf="hasIcon"></b>
-        <ng-content selector="label"></ng-content>
-        <ng-content selector="badge"></ng-content>
-        <ng-content selector="button"></ng-content>
-    </div>
+    <section>
+        <div class="item-list-heading"
+            [class.sticky]="sticky"
+            [class.has-icon]="hasIcon"
+            [class.has-checkbox]="hasCheckbox"
+            [class.has-botonera]="hasBotonera">
+            <ng-content selector="label"></ng-content>
+        </div>
+    </section>
     `
 })
 export class PlexHeadingComponent {
@@ -22,6 +24,7 @@ export class PlexHeadingComponent {
 
     public hasIcon = false;
     public hasCheckbox = false;
+    public hasBotonera = false;
 
     setSticky(value: boolean) {
         this.sticky = value;
@@ -35,6 +38,11 @@ export class PlexHeadingComponent {
 
     setCheckbox(value: boolean) {
         this.hasCheckbox = value;
+        this.ref.detectChanges();
+    }
+
+    setBotonera(value: boolean) {
+        this.hasBotonera = value;
         this.ref.detectChanges();
     }
 }