+++ src/lib/options/options.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, Output, EventEmitter, OnInit } from '@angular/core';
+import { Component, Input, Output, EventEmitter, OnInit, OnChanges, SimpleChanges } from '@angular/core';
 
 
 export interface IPlexOptionsItems {
@@ -10,9 +10,9 @@ export interface IPlexOptionsItems {
     selector: 'plex-options',
     template: `
         <div class="row">
-            <div class="btn-group col">
+            <div class="d-flex col flex-wrap">
                 <ng-container *ngFor="let item of items">
-                    <button class="btn btn-primary btn-sm btn-block m-0" (click)="onOptionsClick(item)" [class.active]="active === item.key">
+                    <button class="btn btn-primary btn-sm option-grow m-0" (click)="onOptionsClick(item)" [class.active]="active === item.key">
                         {{ item.label }}
                     </button>
                 </ng-container>
@@ -20,7 +20,7 @@ export interface IPlexOptionsItems {
         </div>
     `
 })
-export class PlexOptionsComponent implements OnInit {
+export class PlexOptionsComponent implements OnInit, OnChanges {
 
     /**
      * Listado de items a mostrar
@@ -53,8 +53,22 @@ export class PlexOptionsComponent implements OnInit {
         }
     }
 
+    ngOnChanges(changes: SimpleChanges) {
+        if (changes.items) {
+            this.checkKey();
+        }
+    }
+
     onOptionsClick(item: IPlexOptionsItems) {
         this.active = item.key;
         this.activated.emit(item.key);
+    }
+
+    private checkKey() {
+        const isKeyPresent = this.items.some(item => item.key === this.active);
+        if (!isKeyPresent) {
+            this.active = this.items[0].key;
+            this.activated.emit(this.active);
+        }
     }
 }