+++ src/lib/title/title.component.ts
@@ -3,9 +3,9 @@ import { Component, Input, Renderer } from '@angular/core';
 @Component({
     selector: 'plex-title',
     template: `
-        <div class="plex-title d-flex flex-row justify-content-between align-items-center">
+        <div class="plex-title d-flex flex-row justify-content-between align-items-center" responsive>
             <div class="plex-title-label {{ size }}"> {{ titulo }} </div>
-            <div>
+            <div class="title-content">
                 <ng-content></ng-content>
             </div>
         </div>